Support for LOCAL_BUILD option
Added support for the LOCAL_BUILD flag in cmake which sets the version
number of the package to 999.9.9 to avoid confusion with the release
builds.

Also added support for appending the git commit hash to the filename so
that each build result is unique to the commit it was built on, which
further helps with unique versioning.